% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{generate_continuous}
\alias{generate_continuous}
\title{generate_continuous}
\usage{
generate_continuous(p = 5, n1 = 60, n2 = 60, n3 = 60)
}
\arguments{
\item{p}{positive integer; number of variables in the data matrix}

\item{n1}{positive integer; number of individuals in the first interval}

\item{n2}{positive integer; number of individuals in the second interval}

\item{n3}{positive integer; number of individuals in the third interval}
}
\value{
Returns list with the following values:
\enumerate{
\item \code{data}: a \verb{(n1 + n2 + n3) x p} numeric matrix, where the \code{i}-th row is
drawn from a \code{p}-dimensional Gaussian with mean \code{0} and precision matrix
\code{true_precision}
\item \code{covts}: a \verb{(n1 + n2 + n3) x 1} numeric matrix, where the \code{i}-th entry is
the extraneous covariate \code{z_i} for individual \code{i}.
}

The first \code{n1} individuals have \code{z_i} from from a uniform distribution on the
interval \verb{(-3, -1)}.

Individuals \code{n1 + 1} to \code{n1 + n2} have \code{z_i} from from a uniform distribution
on the interval \verb{(-1, 1)}.

Individuals \code{n1 + n2 + 1} to \code{n1 + n2 + n3} have \code{z_i} from a uniform
distribution on the interval \verb{(1, 3)}
\enumerate{
\item \code{true_precision}: \code{list} of \verb{n1 + n2 + n3 p x p} matrices; the \code{i}-th
\code{matrix} is the precision matrix for the \code{i}-th individual
}

All precision matrices have \code{2} on the diagonal and \code{1} in the
\verb{(2, 3)/(3, 2)} position

Individuals in the first interval (\verb{(-3, -1)}) have a \code{1} in the
\verb{(1, 2)/(2, 1)} position, while individuals in the third interval (\verb{(1, 3)})
have a \code{1} in the \verb{(1, 3)/(3, 1)} position.

Individuals in the second interval (\verb{(-1, 1)}) have 2 entries that vary as
a linear function of their extraneous covariate. Let \code{beta = 1/2}. Then,
the \verb{(1, 2)/(2, 1)} position for the \code{i}-th individual in interval 2 is equal
to \code{beta * (1 - z_i)}, while the \verb{(1, 3)/(3, 1)} entry is equal to
\code{beta * (1 + z)}.

Thus, as \code{z_i} approaches \code{-1} from the right, the precision matrix becomes
more similar to the matrix for individuals in interval 1. Similarly, as \code{z_i}
approaches \code{1} from the left, the matrix becomes more similar to the matrix
for individuals in interval \code{3}
}
\description{
function to generate a \code{1}-dimensional extraneous covariate and
\code{p}-dimensional Gaussian data with a precision matrix having entries that
vary as a continuous function of the extraneous covariate
}
