% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{matViz}
\alias{matViz}
\title{matViz}
\usage{
matViz(
  x,
  color1 = "white",
  color2 = "#500000",
  shade = F,
  grid_color = "black",
  incl_val = F,
  prec = 2,
  font_size = 3,
  font_color1 = "black",
  font_color2 = "white",
  font_thres = mean(x)
)
}
\arguments{
\item{x}{\code{matrix}; \code{matrix} to be visualized}

\item{color1}{color; color for low entries. \code{"white"} by default}

\item{color2}{color; color for high entries. \code{"#500000"} by default}

\item{shade}{\code{logical}; if \code{T}, then entries will be shaded according to value
on a gradient ranging from \code{color_low} for the least value to \code{color_high}
for the greatest value}

\item{grid_color}{color; color of grid lines. \code{"black"} by default}

\item{incl_val}{logical; if \code{T}, the value for each entry will be displayed.
\code{F} by default}

\item{prec}{positive integer; number of decimal places to round entries to if
\code{incl_val} is \code{T}. \code{2} by default}

\item{font_size}{positive \code{numeric}; size of font if \code{incl_val} is \code{T}. \code{3}
by default}

\item{font_color1}{color; color of font for low entries if \code{incl_val} is \code{T}.
\code{"black"} by default}

\item{font_color2}{color; color of font for high entries if \code{incl_val} is
\code{T}. \code{"white"} by default}

\item{font_thres}{\code{numeric}; values less than \code{font_thres} will be displayed
in \code{font_color1} if \code{incl_val} is \code{T}. \code{mean(x)} by default}
}
\value{
Returns \code{ggplot2} visualization of \code{matrix}
}
\description{
Create a visualization of a matrix
}
