% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covdepGE_package.R
\docType{package}
\name{covdepGE-package}
\alias{covdepGE-package}
\alias{_PACKAGE}
\title{covdepGE: Covariate Dependent Graph Estimation}
\description{
A weighted psuedo-likelihood approach for modelling the conditional dependence structure of data as a function of an extraneous covariate. Inference is conducted via a block-mean variational approximation
}
\author{
\strong{Maintainer}: Jacob Helwig \email{jacob.a.helwig@tamu.edu}

Authors:
\itemize{
  \item Sutanoy Dasgupta \email{sutanoy@stat.tamu.edu}
  \item Debdeep Pati \email{debdeep@stat.tamu.edu}
}

}
\keyword{internal}
