% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covdepGE_package.R
\docType{package}
\name{covdepGE-package}
\alias{covdepGE-package}
\alias{_PACKAGE}
\title{covdepGE: Covariate Dependent Graph Estimation}
\description{
Assume a \eqn{n} by \eqn{p} data matrix of independent observations, \eqn{X}, and a \eqn{n} by \eqn{p′} matrix of extraneous covariates, \eqn{Z}. The conditional dependence structure of \eqn{X} can be modeled as an undirected graph \eqn{G} such that there is an edge between two variables if, and only if, these variables are dependent on each other given all other variables. Further suppose that the conditional dependence structure of \eqn{X} is not homogeneous across the individuals, and is instead a continuous function of the extraneous covariates \eqn{Z}. Then, this methodology aims to estimate a graph for each of the individuals, possibly unique to the individual, such that similar estimates are made for those who are similar to one another in terms of the extraneous covariate.
}
\author{
\strong{Maintainer}: Jacob Helwig \email{jacob.a.helwig@stat.tamu.edu}

Authors:
\itemize{
  \item Sutanoy Dasgupta \email{sutanoy@stat.tamu.edu}
  \item Debdeep Pati \email{debdeep@stat.tamu.edu}
  \item Patrick Ding \email{patrickding@stat.tamu.edu}
}

}
\keyword{internal}
