% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covdepGE_package.R
\docType{package}
\name{covdepGE-package}
\alias{covdepGE-package}
\alias{_PACKAGE}
\title{covdepGE: Covariate Dependent Graph Estimation}
\description{
A weighted psuedo-likelihood approach for modeling the conditional dependence structure of data as a function of an extraneous covariate. Inference is conducted via a block mean-field variational approximation.
}
\details{
The core function, \code{covdepGE}, uses the weighted psuedo-likelihood
approach to estimate the conditional dependence structure of the data as a
function of an extraneous covariate. Inference is conducted efficiently via
a parallelized block mean-field variational approximation. Three choices
for hyperparameter specification are offered, the default being a hybrid
between model averaging and grid search.

Additionally, the function \code{generateData} returns covariate dependent data
based on the data from the simulation study in (1). The functions
\code{inclusionCurve}, \code{matViz}, and \code{plot.covdepGE} enable visualization of the
estimates returned by the \code{covdepGE} function.
}
\references{
(1) Sutanoy Dasgupta, Peng Zhao, Prasenjit Ghosh, Debdeep Pati, and Bani
Mallick. An approximate Bayesian approach to covariate-dependent graphical
modeling. pages 1–59, 2022.
}
\author{
\strong{Maintainer}: Jacob Helwig \email{jacob.a.helwig@tamu.edu}

Authors:
\itemize{
  \item Sutanoy Dasgupta \email{sutanoy@stat.tamu.edu}
  \item Peng Zhao \email{pzhao@stat.tamu.edu}
  \item Bani Mallick \email{bmallick@stat.tamu.edu}
  \item Debdeep Pati \email{debdeep@stat.tamu.edu}
}

}
