% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covdepGE.R
\name{covdepGE}
\alias{covdepGE}
\title{covdepGE}
\usage{
covdepGE(
  data_mat,
  Z,
  tau = 0.1,
  sigmavec = c(0.01, 0.05, 0.1, 0.5, 1, 3, 7, 10),
  tolerance = 1e-09,
  max_iter = 100
)
}
\arguments{
\item{max_iter}{}
}
\value{

}
\description{
covdepGE
}
