% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covdepGE_main.R
\name{covdepGE}
\alias{covdepGE}
\title{Title}
\usage{
covdepGE(
  data_mat,
  Z,
  tau = 0.1,
  kde = T,
  alpha = 0.2,
  mu = 0,
  sigmasq = 0.5,
  sigmabetasq_vec = NULL,
  var_min = 0.01,
  var_max = 10,
  n_sigma = 8,
  pi_vec = 0.2,
  norm = 2,
  scale = T,
  tolerance = 1e-09,
  max_iter = 100,
  edge_threshold = 0.5,
  sym_method = "mean",
  print_time = F,
  CS = F
)
}
\arguments{
\item{CS}{}
}
\value{

}
\description{
Title
}
