% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covdepGE_main.R
\name{covdepGE}
\alias{covdepGE}
\title{Title}
\usage{
covdepGE(
  data_mat,
  Z,
  tau = 0.1,
  alpha = 0.2,
  mu = 0,
  sigmavec = c(0.01, 0.05, 0.1, 0.5, 1, 3, 7, 10),
  pi_vec = seq(0.1, 0.9, 0.1),
  scale = T,
  tolerance = 1e-09,
  max_iter = 100,
  edge_threshold = 0.5,
  print_time = F
)
}
\arguments{
\item{print_time}{}
}
\value{

}
\description{
Title
}
