% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_covdepGE.R
\name{gg_adjMat}
\alias{gg_adjMat}
\title{Title}
\usage{
gg_adjMat(
  out,
  l,
  prob_shade = T,
  color0 = "white",
  color1 = "#500000",
  grid_color = "black",
  incl_probs = T,
  prob_prec = 2,
  font_size = 3,
  font_color0 = "black",
  font_color1 = "white"
)
}
\arguments{
\item{font_color1}{}
}
\value{

}
\description{
Title
}
