% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_covdepGE.R
\name{gg_inclusionCurve}
\alias{gg_inclusionCurve}
\title{Title}
\usage{
gg_inclusionCurve(
  out,
  col_idx1,
  col_idx2,
  line_type = "solid",
  line_size = 0.5,
  line_color = "black",
  point_shape = 21,
  point_size = 1.5,
  point_color = "#500000",
  point_fill = "white",
  sort = F
)
}
\arguments{
\item{sort}{}
}
\value{

}
\description{
Title
}
