% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{inclusionCurve}
\alias{inclusionCurve}
\title{inclusionCurve}
\usage{
inclusionCurve(
  out,
  col_idx1,
  col_idx2,
  line_type = "solid",
  line_size = 0.5,
  line_color = "black",
  point_shape = 21,
  point_size = 1.5,
  point_color = "#500000",
  point_fill = "white"
)
}
\arguments{
\item{out}{object of class \code{covdepGE}; return of \code{covdepGE} function}

\item{col_idx1}{integer in \verb{[1, p]}; column index of the first variable}

\item{col_idx2}{integer in \verb{[1, p]}; column index of the second variable}

\item{line_type}{linetype; \code{ggplot2} line type to interpolate the
probabilities. \code{"solid"} by default}

\item{line_size}{positive \code{numeric}; thickness of the interpolating line.
\code{0.5} by default}

\item{line_color}{color; color of interpolating line. \code{"black"} by default}

\item{point_shape}{shape; shape of the points denoting individual-specific
inclusion probabilities; \code{21} by default}

\item{point_size}{positive \code{numeric}; size of probability points. \code{1.5} by
default}

\item{point_color}{color; color of probability points. \code{"#500000"} by default}

\item{point_fill}{color; fill of probability points. Only applies to select
shapes. \code{"white"} by default}
}
\value{
Returns \code{ggplot2} visualization of inclusion probability curve
}
\description{
Create a visualization of the probabilities of inclusion of an
edge between two variables across all \code{n} individuals
}
