% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{matViz}
\alias{matViz}
\title{matViz}
\usage{
matViz(
  x,
  color1 = "white",
  color2 = "#500000",
  grid_color = "black",
  incl_val = F,
  prec = 2,
  font_size = 3,
  font_color1 = "black",
  font_color2 = "white",
  font_thres = mean(x)
)
}
\arguments{
\item{x}{\code{matrix}; \code{matrix} to be visualized}

\item{color1}{color; color for low entries. \code{"white"} by default}

\item{color2}{color; color for high entries. \code{"#500000"} by default}

\item{grid_color}{color; color of grid lines. \code{"black"} by default}

\item{incl_val}{logical; if \code{T}, the value for each entry will be displayed.
\code{F} by default}

\item{prec}{positive \code{integer}; number of decimal places to round entries to if
\code{incl_val} is \code{T}. \code{2} by default}

\item{font_size}{positive \code{numeric}; size of font if \code{incl_val} is \code{T}. \code{3}
by default}

\item{font_color1}{color; color of font for low entries if \code{incl_val} is \code{T}.
\code{"black"} by default}

\item{font_color2}{color; color of font for high entries if \code{incl_val} is
\code{T}. \code{"white"} by default}

\item{font_thres}{\code{numeric}; values less than \code{font_thres} will be displayed
in \code{font_color1} if \code{incl_val} is \code{T}. \code{mean(x)} by default}
}
\value{
Returns \code{ggplot2} visualization of \code{matrix}
}
\description{
Create a visualization of a \code{matrix}
}
\examples{

library(ggplot2)

# get the data
set.seed(1)
data <- generateData()
X <- data$data
Z <- data$covts
interval <- data$interval
prec <- data$true_precision

# get overall and within interval sample sizes
n <- nrow(X)
n1 <- sum(interval == 1)
n2 <- sum(interval == 2)

# visualize the distribution of the extraneous covariate
ggplot(data.frame(Z = Z, interval = as.factor(interval))) +
geom_histogram(aes(Z, fill = interval), color = "black", bins = n \%/\% 5)

# visualize the true precision matrices in each of the intervals

# interval 1
matViz(prec[[1]], incl_val = TRUE) +
ggtitle("True precision matrix, interval 1")

# interval 2 (varies continuously with Z)
int2_mats <- prec[interval == 2]
int2_inds <- c(5, n2 \%/\% 2, n2 - 5)
lapply(int2_inds, function(j) matViz(int2_mats[[j]], incl_val = TRUE) +
ggtitle(paste("True precision matrix, interval 2, individual", j)))

# interval 3
matViz(prec[[length(prec)]], incl_val = TRUE) +
ggtitle("True precision matrix, interval 3")

# fit the model and visualize the estimated precision matrices
(out <- covdepGE(X, Z))
plot(out)

# visualize the inclusion probabilities for variables (1, 3) and (1, 2)
inclusionCurve(out, 1, 2)
inclusionCurve(out, 1, 3)

}
