% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.covdepGE}
\alias{plot.covdepGE}
\title{plot.covdepGE}
\usage{
\method{plot}{covdepGE}(x, graph_colors = NULL, title_sum = T, ...)
}
\arguments{
\item{x}{object of class \code{covdepGE}; return of \code{covdepGE} function}

\item{graph_colors}{\code{NULL} OR \code{vector} of length \code{g}; \code{g} is the number of
unique graphs from \code{x}. The \code{v}-th element is the color for the \code{v}-th unique
graph. If \code{NULL}:

\code{graph_colors <- rep("#500000", length(out$unique_graphs))}

\code{NULL} by default}

\item{title_sum}{\code{logical}; whether the indices of the individuals
corresponding to the graph should be included in the title. \code{T} by default}

\item{...}{additional arguments will be ignored}
}
\value{
Returns list of \code{ggplot2} visualizations of unique adjacency matrices
estimated by \code{covdepGE}
}
\description{
Given the return value of \code{covdepGE} function, create a list of
visualizations of the adjacency matrix for each of the unique graphs
}
